% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceterisParibusD3.R
\name{ceterisParibusD3}
\alias{ceterisParibusD3}
\title{Create interactive D3-based Ceteris Paribus Explanations Plots}
\usage{
ceterisParibusD3(model, ..., selected_variables = NULL, color = NULL,
  width = NULL, height = NULL, no_colors = 3,
  is_categorical_ordered = FALSE, size_rugs = NULL,
  alpha_rugs = NULL, color_rugs = NULL, color_points = NULL,
  color_residuals = NULL, color_pdps = NULL, alpha_residuals = NULL,
  alpha_points = NULL, alpha_ices = NULL, alpha_pdps = NULL,
  size_points = NULL, size_residuals = NULL, size_ices = NULL,
  size_pdps = NULL, show_profiles = TRUE, show_observations = TRUE,
  show_rugs = NULL, show_residuals = NULL, aggregate_profiles = NULL,
  font_size_titles = NULL, font_size_legend = NULL,
  font_size_axes = NULL, font_size_tootlips = NULL,
  font_size_table = NULL, add_table = NULL,
  font_size_plot_title = NULL, plot_title = NULL, yaxis_title = NULL,
  auto_resize = TRUE)
}
\arguments{
\item{model}{a ceteris paribus explainer produced with function `ceteris_paribus()` from ceterisParibus package}

\item{...}{other explainers that shall be plotted together}

\item{selected_variables}{if not NULL then only `selected_variables` will be presented}

\item{color}{a character. Either name of a color or name of a variable that should be used for coloring. If color is a categorical variable, it should have max. 9 categories. Use '_label_' to color by model type}

\item{width}{a numeric. Width (in px) of the whole plot}

\item{height}{a numeric. Heigth (in px) of the whole plot}

\item{no_colors}{a numeric. Number of colors in legend for sequential scales (currently available: between 1 and 9)}

\item{is_categorical_ordered}{a logical. If TRUE categorical variables values on x axis will be sorted alphabetically}

\item{size_rugs}{a numeric. Size of rugs to be plotted, between 0 and 1}

\item{alpha_rugs}{a numeric between 0 and 1. Opacity of rug lines}

\item{color_rugs}{a character. Name of a color. If NULL elements are plotted according to 'color' arguments}

\item{color_points}{a character. Name of a color. If NULL elements are plotted according to 'color' arguments}

\item{color_residuals}{a character. Name of a color. If NULL elements are plotted according to 'color' arguments}

\item{color_pdps}{a character. Name of a color.}

\item{alpha_residuals}{a numeric between 0 and 1. Opacity of residuals}

\item{alpha_points}{a numeric between 0 and 1. Opacity of points}

\item{alpha_ices}{a numeric between 0 and 1. Opacity of ICE lines}

\item{alpha_pdps}{a numeric between 0 and 1. Opacity of PDP lines}

\item{size_points}{a numeric. Size of points to be plotted}

\item{size_residuals}{a numeric. Size of residuals (lines and points) to be plotted}

\item{size_ices}{a numeric. Size of ICE lines to be plotted}

\item{size_pdps}{a numeric. Size of PDP lines to be plotted}

\item{show_profiles}{a logical. If TRUE then individual profiles will be plotted}

\item{show_observations}{a logical. If TRUE then individual observations will be marked as points}

\item{show_rugs}{a logical. If TRUE then individual observations will be marked as rugs}

\item{show_residuals}{a logical. If TRUE then residuals will be plotted as a line ended with a point}

\item{aggregate_profiles}{a character. Either 'mean' or 'median'. If not NULL then profiles will be aggregated according to chosen metric and the aggregate profiles will be plotted}

\item{font_size_titles}{a numeric. Font size in px of subplots titles}

\item{font_size_legend}{a numeric. Font size in px of texts in legend}

\item{font_size_axes}{a numeric. Font size in px of texts on axes}

\item{font_size_tootlips}{a numeric. Font size in px of texts in tooltip}

\item{font_size_table}{a numeric. Font size in px of texts in table}

\item{add_table}{a logical. If TRUE then table will be plotted, default is TRUE}

\item{font_size_plot_title}{a numeric. Font size in px of plot main title}

\item{plot_title}{a character. Main title of the plot}

\item{yaxis_title}{a character. Vertical (y) axis title of the plot}

\item{auto_resize}{a logical. If FALSE plot's elements (like fonts or lines) won't be automatically resized when changinh size of the window.}
}
\value{
a ceterisParibusD3 object
}
\description{
Function 'ceterisParibusD3' plots interactive version of Ceteris Paribus Plots available in ceterisParibus package.
Various parameters help to decide what (profiles, aggregated profiles, points or rugs) and how it should be plotted.
}
\examples{

 \dontrun{
library("DALEX")
library("ceterisParibusD3")
library("ceterisParibus")
library("randomForest")
set.seed(59)


apartments_rf_model <- randomForest(m2.price ~ construction.year + surface + floor +
                                     no.rooms + district,
                                   data = apartments)

explainer_rf <- explain(apartments_rf_model,
                       data = apartmentsTest[,2:6],
                       y = apartmentsTest$m2.price)

apartments_A <- apartmentsTest[958,]
cp_rf_A <- ceteris_paribus(explainer_rf, apartments_A, y = apartments_A$m2.price)

# plot in ceterisParibus package
plot(cp_rf_A, show_profiles = TRUE, show_observations = TRUE,
    selected_variables = c("surface","construction.year"))

# interactive version from ceterisParibusD3 package
ceterisParibusD3(cp_rf_A, show_profiles = TRUE, show_observations = TRUE,
                selected_variables = c("surface","construction.year"), add_table = FALSE)

}

}
